Identities Example - Merge TrueIDs based on CSV
This is the initial example showcasing how to merge two TrueIdentity records together based on two different TrueIdentity Login Identifiers.

In this example the Identifier that is listed as the Privileged_SamAccountName provides the lookup into the TrueIdentity record that will be merged into the TrueIdentity Record containing the SamAccountName.  The script has the appropriate logic to prevent attempting to merge identifiers that have previously been merged.

The Example also supports leveraging the produced output as an input, granting the ability to review the merge results and perform any manual modifications necissary before re-running the merger task.